Add image info
fix CLI
