+++ TagCloud/IImageInfo.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+
+namespace TagCloud
+{
+    public interface IImageInfo
+    {
+        List<Tuple<String, Rectangle>> GetTags(string filename, int canvasHeight);
+    }
+}
