+++ TagCloud/Visualizer.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using TagCloud.Layout;
@@ -8,33 +7,28 @@ namespace TagCloud
 {
     public class Visualizer: IVisualizer
     {
-        private IFrequencyAnalyzer FrequencyAnalyzer;
-        private ILayouter Layouter;
         private ICanvas Canvas;
         private IPathCreater Creater;
-        public Visualizer(IFrequencyAnalyzer frequencyAnalyzer, ILayouter layouter, ICanvas canvas, IPathCreater pathCreator)
+        private IImageInfo ImageInfo;
+        public Visualizer(ICanvas canvas, IPathCreater pathCreator, IImageInfo imageInfo)
         {
             //TODO: add fontFamily and coloring algoritm
-            FrequencyAnalyzer = frequencyAnalyzer;
-            Layouter = layouter;
             Canvas = canvas;
             Creater = pathCreator;
+            ImageInfo = imageInfo;
         }
 
         public void Visualize(string filename)
         {
-            var frequencies = FrequencyAnalyzer.GetFrequencyDictionary(filename);
             var bitmap = new Bitmap(Canvas.Width, Canvas.Height);
             var graphics = Graphics.FromImage(bitmap);
-            
-            var orderedPairs = frequencies.OrderByDescending(pair => pair.Value);
-            foreach (var pair in orderedPairs)
+
+            foreach (var pair in ImageInfo.GetTags(filename, Canvas.Height))
             {
-                var height = (int)Math.Round(Canvas.Height * pair.Value);
-                var width = (int)Math.Round((double)height * pair.Key.Length / 2);
-                var rectangle = Layouter.PutNextRectangle(new Size(width, height));
+                var rectangle = pair.Item2;
                 DrawAndFillRectangle(graphics, rectangle);
-                graphics.DrawString(pair.Key, new Font("Arial", height/2), new SolidBrush(Color.Black), rectangle);
+                graphics.DrawString(pair.Item1, new Font("Arial",rectangle.Height/2),
+                    new SolidBrush(Color.Black), rectangle);
             }
             
             bitmap.Save(Creater.GetNewPngPath());
